Solved: Koko Eating Bananas (#907)

+++ problems/0907-koko-eating-bananas/solution.cpp
@@ -2,14 +2,14 @@ class Solution {
 public:
     int Maxelen(vector<int> piles){
         int n = piles.size();
-        int ans = piles[0];
+        long long ans = piles[0];
         for(int i = 1; i<n;i++){
             if(piles[i] > ans)  ans = piles[i];
         }
         return ans;
     }
-    int func(vector<int> piles, int mid, int n){
-        int totalhrs = 0;
+    long long func(vector<int> piles, int mid, int n){
+        long long totalhrs = 0;
         for(int i = 0; i < n; i++){
             totalhrs += ceil((double)piles[i]/mid);
         }
@@ -18,12 +18,12 @@ public:
 
     int minEatingSpeed(vector<int>& piles, int h) {
         int n = piles.size();
-        int low = 1, high = Maxelen(piles);
+        long long low = 1, high = Maxelen(piles);
         while(low <= high){
-            int mid = (low + high)/2;
-            int k = func(piles, mid, n);
+            long long mid = (low + high)/2;
+            long long k = func(piles, mid, n);
             if(k <= h){
-                                high = mid - 1;
+                high = mid - 1;
             }
             else{
                 low = mid + 1;
